+++ templates/shared_services/admin-vm/terraform/admin-jumpbox.tf
@@ -45,12 +45,28 @@ resource "azurerm_windows_virtual_machine" "jumpbox" {
   }
 
   os_disk {
-    name                 = "vm-dsk-${var.tre_id}"
-    caching              = "ReadWrite"
-    storage_account_type = "Standard_LRS"
+    name                   = "vm-dsk-${var.tre_id}"
+    caching                = "ReadWrite"
+    storage_account_type   = "Standard_LRS"
+    disk_encryption_set_id = var.enable_cmk_encryption ? azurerm_disk_encryption_set.jumpbox_disk_encryption[0].id : null
   }
 
   lifecycle { ignore_changes = [tags] }
+}
+
+resource "azurerm_disk_encryption_set" "jumpbox_disk_encryption" {
+  count                     = var.enable_cmk_encryption ? 1 : 0
+  name                      = "disk-encryption-jumpbox-${var.tre_id}-${var.tre_resource_id}"
+  location                  = data.azurerm_resource_group.rg.location
+  resource_group_name       = data.azurerm_resource_group.rg.name
+  key_vault_key_id          = data.azurerm_key_vault_key.tre_encryption_key[0].versionless_id
+  encryption_type           = "EncryptionAtRestWithPlatformAndCustomerKeys"
+  auto_key_rotation_enabled = true
+
+  identity {
+    type         = "UserAssigned"
+    identity_ids = [data.azurerm_user_assigned_identity.tre_encryption_identity[0].id]
+  }
 }
 
 resource "azurerm_key_vault_secret" "jumpbox_credentials" {