+++ templates/shared_services/certs/terraform/variables.tf
@@ -14,3 +14,12 @@ variable "tre_resource_id" {
   type        = string
   description = "Resource ID"
 }
+
+variable "enable_cmk_encryption" {
+  type    = bool
+  default = false
+}
+
+variable "key_store_id" {
+  type = string
+}