+++ templates/shared_services/cyclecloud/terraform/variables.tf
@@ -7,3 +7,10 @@ variable "tre_resource_id" {
 variable "arm_environment" {
   type = string
 }
+variable "enable_cmk_encryption" {
+  type    = bool
+  default = false
+}
+variable "key_store_id" {
+  type = string
+}