+++ templates/workspace_services/azureml/terraform/storage.tf
@@ -4,16 +4,34 @@ resource "azurerm_storage_account" "aml" {
   resource_group_name              = data.azurerm_resource_group.ws.name
   account_tier                     = "Standard"
   account_replication_type         = "GRS"
+  table_encryption_key_type        = var.enable_cmk_encryption ? "Account" : "Service"
+  queue_encryption_key_type        = var.enable_cmk_encryption ? "Account" : "Service"
   cross_tenant_replication_enabled = false
   tags                             = local.tre_workspace_service_tags
   network_rules {
     default_action = "Deny"
   }
 
+  dynamic "identity" {
+    for_each = var.enable_cmk_encryption ? [1] : []
+    content {
+      type         = "UserAssigned"
+      identity_ids = [data.azurerm_user_assigned_identity.ws_encryption_identity[0].id]
+    }
+  }
+
   # changing this value is destructive, hence attribute is in lifecycle.ignore_changes block below
   infrastructure_encryption_enabled = true
 
   lifecycle { ignore_changes = [infrastructure_encryption_enabled, tags] }
+}
+
+resource "azurerm_storage_account_customer_managed_key" "aml_stg_encryption" {
+  count                     = var.enable_cmk_encryption ? 1 : 0
+  storage_account_id        = azurerm_storage_account.aml.id
+  key_vault_id              = var.key_store_id
+  key_name                  = local.cmk_name
+  user_assigned_identity_id = data.azurerm_user_assigned_identity.ws_encryption_identity[0].id
 }
 
 data "azurerm_private_dns_zone" "blobcore" {
@@ -48,7 +66,6 @@ resource "azurerm_private_endpoint" "blobpe" {
 
 
 }
-
 
 resource "azurerm_private_endpoint" "filepe" {
   name                = "pe-file-${local.storage_name}"