+++ templates/workspace_services/guacamole/user_resources/guacamole-azure-windowsvm/terraform/data.tf
@@ -42,3 +42,15 @@ data "azurerm_storage_share" "shared_storage" {
   name                 = var.shared_storage_name
   storage_account_name = data.azurerm_storage_account.stg.name
 }
+
+data "azurerm_key_vault_key" "ws_encryption_key" {
+  count        = var.enable_cmk_encryption ? 1 : 0
+  name         = local.cmk_name
+  key_vault_id = var.key_store_id
+}
+
+data "azurerm_user_assigned_identity" "ws_encryption_identity" {
+  count               = var.enable_cmk_encryption ? 1 : 0
+  name                = local.encryption_identity_name
+  resource_group_name = data.azurerm_resource_group.ws.name
+}