+++ templates/workspaces/base/terraform/airlock/storage_accounts.tf
@@ -5,6 +5,8 @@ resource "azurerm_storage_account" "sa_import_approved" {
   resource_group_name              = var.ws_resource_group_name
   account_tier                     = "Standard"
   account_replication_type         = "LRS"
+  table_encryption_key_type        = var.enable_cmk_encryption ? "Account" : "Service"
+  queue_encryption_key_type        = var.enable_cmk_encryption ? "Account" : "Service"
   allow_nested_items_to_be_public  = false
   cross_tenant_replication_enabled = false
 
@@ -68,6 +70,8 @@ resource "azurerm_storage_account" "sa_export_internal" {
   resource_group_name              = var.ws_resource_group_name
   account_tier                     = "Standard"
   account_replication_type         = "LRS"
+  table_encryption_key_type        = var.enable_cmk_encryption ? "Account" : "Service"
+  queue_encryption_key_type        = var.enable_cmk_encryption ? "Account" : "Service"
   allow_nested_items_to_be_public  = false
   cross_tenant_replication_enabled = false
 
@@ -131,6 +135,8 @@ resource "azurerm_storage_account" "sa_export_inprogress" {
   resource_group_name              = var.ws_resource_group_name
   account_tier                     = "Standard"
   account_replication_type         = "LRS"
+  table_encryption_key_type        = var.enable_cmk_encryption ? "Account" : "Service"
+  queue_encryption_key_type        = var.enable_cmk_encryption ? "Account" : "Service"
   allow_nested_items_to_be_public  = false
   cross_tenant_replication_enabled = false
 
@@ -201,6 +207,8 @@ resource "azurerm_storage_account" "sa_export_rejected" {
   resource_group_name              = var.ws_resource_group_name
   account_tier                     = "Standard"
   account_replication_type         = "LRS"
+  table_encryption_key_type        = var.enable_cmk_encryption ? "Account" : "Service"
+  queue_encryption_key_type        = var.enable_cmk_encryption ? "Account" : "Service"
   allow_nested_items_to_be_public  = false
   cross_tenant_replication_enabled = false
 
@@ -264,6 +272,8 @@ resource "azurerm_storage_account" "sa_export_blocked" {
   resource_group_name              = var.ws_resource_group_name
   account_tier                     = "Standard"
   account_replication_type         = "LRS"
+  table_encryption_key_type        = var.enable_cmk_encryption ? "Account" : "Service"
+  queue_encryption_key_type        = var.enable_cmk_encryption ? "Account" : "Service"
   allow_nested_items_to_be_public  = false
   cross_tenant_replication_enabled = false